perf(hosted-runner): defer device-sync maintenance off the cold boot path
The hosted runner's cold-start nodeStartupMs pays for every module in the
static boot closure. maintenance.ts imported @murphai/device-syncd
{config,service,registry} at module scope, dragging device-syncd,
@murphai/importers, and @junction-api/sdk (~987KB static, ~912KB of it in
the entry chunk) onto the boot path — yet nothing from those imports runs
at module-evaluation time. The registry and service are constructed only
inside the per-wake/idle device-sync lanes, seconds to minutes after boot.

Same shape as #397's conversation-connector deferral, one layer up:

- Split the device-sync maintenance surface (runHostedDeviceSyncPass,
  runHostedDeviceSyncWakeLane, resolveHostedDeviceSyncNextWakeAt and their
  private helpers) out of maintenance.ts into device-sync-maintenance.ts,
  moved verbatim. The assistant-automation lanes that
  workspace-assistant-phase.ts needs eagerly stay in maintenance.ts, which
  now has zero device-syncd imports.
- Load the new module behind the house cached-dynamic-import pattern
  (device-sync-maintenance-import.ts): promise cached, cache reset on
  rejection so retry works, rejection wrapped in a typed load error. A
  load failure surfaces loudly as a distinct
  device-sync.module_load_failed runtime log (registered in the
  hosted-execution log contract), never a silent no-op of the wake lane.
- Narrow the remaining connect/reconnect prompt helpers to the light
  @murphai/device-syncd/connect-config subpath. No production caller of
  resolveHostedDeviceSyncNextWakeAt sits on the boot path, so it lives in
  the lazy module.
- Extend the boot-closure guard with suffix markers for
  device-syncd/{service,registry,providers}, importers, and the Junction
  SDK (light device-syncd subpaths legitimately stay static), and ratchet
  the entry baseline down to the new measured 1,267,937B.

Runner entry chunk 2,288,516B -> 1,267,937B (-45%); static boot closure
~7.76MB -> 6,809,980B; total ~8.2MB unchanged (the code moves to a lazy
chunk loaded on the first device-sync lane run).

Verified: real assembly passes at the new baseline (guard + ratchet +
lazy-chunk boot probe); runner-bundle vitest 21 passed incl. new marker
coverage in both staged and workspace path shapes; assistant-runtime
maintenance/device-sync/events/phase + new loader tests 270 passed;
assistant-runtime, hosted-execution, cloudflare, and repo-wide typechecks.
