+++ apps/cloudflare/scripts/runner-bundle/bundle-entrypoint.ts
@@ -29,7 +29,7 @@ export const RUNNER_ENTRYPOINT_BUNDLE_DIRECTORY_NAME = "dist-bundled";
 // Byte budgets over the esbuild metafile so import-graph creep in the boot
 // surface fails the assembly instead of silently regressing cold start.
 // Baselines measured from the real assembled bundle on 2026-07-06: total
-// 8,201,281B across 34 chunks, entry container-entrypoint.js 2,288,516B.
+// 8,223,286B across 42 chunks, entry container-entrypoint.js 1,267,937B.
 //
 // The entry chunk gates cold-start parse, so it is ratcheted, not given
 // headroom: the guard holds it to the measured baseline plus a tight noise
@@ -44,20 +44,25 @@ export const RUNNER_ENTRYPOINT_BUNDLE_DIRECTORY_NAME = "dist-bundled";
 // too if total creep becomes the concern. Investigate the listed largest
 // inputs before raising either.
 const RUNNER_ENTRYPOINT_BUNDLE_TOTAL_BYTES_BUDGET = 9_300_000;
-const RUNNER_ENTRYPOINT_BUNDLE_ENTRY_BASELINE_BYTES = 2_288_516;
+const RUNNER_ENTRYPOINT_BUNDLE_ENTRY_BASELINE_BYTES = 1_267_937;
 // Noise band above the baseline before the ratchet trips (~2%): absorbs
 // content-hash and minifier jitter without letting real boot-path weight land
 // silently. Keep it tight; it is a tolerance for noise, not feature headroom.
 const RUNNER_ENTRYPOINT_BUNDLE_ENTRY_TOLERANCE_BYTES = 48_000;
-// The inboxd markers are path suffixes, not node_modules-anchored: workspace
-// package inputs appear as `node_modules/@murphai/inboxd/dist/...` in the
-// staged production assembly but as `packages/inboxd/dist/...` when bundling
+// The @murphai package markers are path suffixes, not node_modules-anchored:
+// workspace package inputs appear as `node_modules/@murphai/*/dist/...` in
+// the staged production assembly but as `packages/*/dist/...` when bundling
 // straight from the repo checkout, and the guard must bite in both shapes.
 const RUNNER_ENTRYPOINT_FORBIDDEN_BOOT_INPUT_MARKERS = [
   "node_modules/grammy/",
   "node_modules/node-fetch/",
   "/inboxd/dist/connectors/hosted-conversation.js",
   "/inboxd/dist/connectors/telegram/connector.js",
+  "/device-syncd/dist/service.js",
+  "/device-syncd/dist/registry.js",
+  "/device-syncd/dist/providers/",
+  "/importers/dist/",
+  "node_modules/@junction-api/sdk/",
 ] as const;
 
 export async function bundleRunnerContainerEntrypoint(
